# kaushaldrishti/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KaushalDrishti API",
    "description": "AI-enabled Labour Market Intelligence System (LMIS) for MSDE. Provides district-level labour demand indices, supply forecasts, gap analysis, early warnings, and policy scenario simulation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/scenarios": {
      "post": {
        "tags": [
          "scenario"
        ],
        "summary": "Create Scenario",
        "description": "Simulate a policy intervention (e.g. +15% seat allocation, new centre, completion bump).\nReturns monthly baseline vs intervention time-series and identifies the gap closure cycle.",
        "operationId": "create_scenario_api_v1_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenarios/{scenario_id}": {
      "get": {
        "tags": [
          "scenario"
        ],
        "summary": "Get Scenario",
        "description": "Retrieve a previously simulated policy scenario by ID.",
        "operationId": "get_scenario_api_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scenario run UUID",
              "title": "Scenario Id"
            },
            "description": "Scenario run UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/lineage/{cell_id}": {
      "get": {
        "tags": [
          "lineage"
        ],
        "summary": "Get Lineage",
        "description": "Retrieve full data lineage and provenance tree for a district-trade cell.\nTraces raw source feeds, weights, gate pass status, reliability r_k, and fallback level.",
        "operationId": "get_lineage_api_v1_lineage__cell_id__get",
        "parameters": [
          {
            "name": "cell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Cell ID in format '{district_id}_{trade_id}' or integer ID",
              "title": "Cell Id"
            },
            "description": "Cell ID in format '{district_id}_{trade_id}' or integer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvidenceLineageItem": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "contribution": {
            "type": "number",
            "title": "Contribution"
          },
          "gate_pass": {
            "type": "boolean",
            "title": "Gate Pass"
          },
          "freshness_days": {
            "type": "number",
            "title": "Freshness Days"
          },
          "reliability_r_k": {
            "type": "number",
            "title": "Reliability R K"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "source_name",
          "source_type",
          "weight",
          "contribution",
          "gate_pass",
          "freshness_days",
          "reliability_r_k"
        ],
        "title": "EvidenceLineageItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LineageResponse": {
        "properties": {
          "cell_id": {
            "type": "string",
            "title": "Cell Id"
          },
          "origin_month": {
            "type": "string",
            "title": "Origin Month"
          },
          "district_id": {
            "type": "integer",
            "title": "District Id"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "state_code": {
            "type": "string",
            "title": "State Code"
          },
          "trade_id": {
            "type": "integer",
            "title": "Trade Id"
          },
          "trade_name": {
            "type": "string",
            "title": "Trade Name"
          },
          "fallback_level": {
            "type": "integer",
            "title": "Fallback Level"
          },
          "fallback_description": {
            "type": "string",
            "title": "Fallback Description"
          },
          "data_mode": {
            "type": "string",
            "title": "Data Mode"
          },
          "latent_mean_m": {
            "type": "number",
            "title": "Latent Mean M"
          },
          "latent_variance_p": {
            "type": "number",
            "title": "Latent Variance P"
          },
          "evidence_units": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLineageItem"
            },
            "type": "array",
            "title": "Evidence Units"
          },
          "pipeline_stages": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Pipeline Stages"
          }
        },
        "type": "object",
        "required": [
          "cell_id",
          "origin_month",
          "district_id",
          "district_name",
          "state_code",
          "trade_id",
          "trade_name",
          "fallback_level",
          "fallback_description",
          "data_mode",
          "latent_mean_m",
          "latent_variance_p",
          "evidence_units",
          "pipeline_stages"
        ],
        "title": "LineageResponse"
      },
...
      "ScenarioRequest": {
        "properties": {
          "district_id": {
            "type": "integer",
            "title": "District Id",
            "description": "Target district ID"
          },
          "trade_id": {
            "type": "integer",
            "title": "Trade Id",
            "description": "Target trade ID"
          },
          "seat_delta_pct": {
            "type": "number",
            "maximum": 0.5,
            "minimum": -0.5,
            "title": "Seat Delta Pct",
            "description": "Seat allocation change % (-0.50 to +0.50)",
            "default": 0.15
          },
          "completion_delta_pp": {
            "type": "number",
            "maximum": 0.3,
            "minimum": -0.3,
            "title": "Completion Delta Pp",
            "description": "Completion rate delta in percentage points",
            "default": 0.0
          },
          "new_centre_capacity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "New Centre Capacity",
            "description": "Additional capacity from a newly sanctioned training centre",
            "default": 0
          },
          "new_centre_build_lag_months": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 0.0,
            "title": "New Centre Build Lag Months",
            "description": "Months before new centre becomes operational",
            "default": 6
          },
          "demand_case": {
            "type": "string",
            "title": "Demand Case",
            "description": "'base', 'high', or 'low'",
            "default": "base"
          },
          "start_month": {
            "type": "string",
            "title": "Start Month",
            "description": "Intervention start cycle (YYYY-MM)",
            "default": "2025-01"
          }
        },
        "type": "object",
        "required": [
          "district_id",
          "trade_id"
        ],
        "title": "ScenarioRequest"
      },
      "ScenarioResponse": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "district_id": {
            "type": "integer",
            "title": "District Id"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "trade_id": {
            "type": "integer",
            "title": "Trade Id"
          },
          "trade_name": {
            "type": "string",
            "title": "Trade Name"
          },
          "parameters": {
            "type": "object",
            "title": "Parameters"
          },
          "summary": {
            "type": "object",
            "title": "Summary"
          },
          "series": {
            "type": "object",
            "title": "Series"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          }
        },
        "type": "object",
        "required": [
          "scenario_id",
          "created_at",
          "district_id",
          "district_name",
          "trade_id",
          "trade_name",
          "parameters",
          "summary",
          "series",
          "assumptions"
        ],
        "title": "ScenarioResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}